demo(android): unify multi-model + scene-gallery into model-viewer (#2239 Batch 5)
Consolidates the retired `multi-model` and `scene-gallery` demos into the
existing flagship `model-viewer` entry with a 3-way segmented-button toggle
(mirrors #2239 Batch 1/2/3/4):

- **Single Model** (default) — the canonical full-screen model viewer:
  bundled hero helmet, auto-fit hero orbit, optional Sketchfab "Surprise me"
  stream. (The original `model-viewer` demo.)
- **Multi-Model** — themed "Park" scene from 4 streamed glTF assets with
  per-model visibility chips + spin toggle. (Formerly `multi-model`.)
- **Gallery** — chip-picked themed Sketchfab CC-BY models on an orbit.
  (Formerly `scene-gallery`.)

The `model-viewer` id AND its `ModelViewerDemo.kt` file are preserved verbatim
because they are the flagship example referenced across docs (llms.txt,
agents/sceneview cheatsheet + recipes) — minimal ripple. Each sub-mode owns
its own rememberEngine + loaders + SceneView so switching tabs tears down the
inactive section completely (prevents resource leaks). All QA-day / #issue
rationale comments preserved from the source demos.

Retired deep-link ids (`multi-model`, `scene-gallery`) keep working via
DeepLinkRouter.DEMO_ID_ALIASES (and the RootScreen Explore-tab curated list
repoints the retired `multi-model` id to the live `materials` umbrella so the
carousel keeps its 6-card diversity). Their fragments + strings folded into
the unified `model-viewer` demo; llms.txt demos block + GeneratedDemos.kt
regenerated via collate-demos.sh. The hand-written Picker-pattern prose in
llms.txt repointed from `MultiModelDemo` to `ModelViewerDemo`.

Tests: DemoInteractionTest.modelViewer_allTabs taps all 3 tabs;
DemoRenderingScreenshotTest.modelViewerDemo_default_state kept (golden
modelviewer_default). Stale multimodel_default.png golden + its
multiModelDemo_default_state test removed.
