+++ changelog.d/2239-batch-5-scene.md
@@ -1,0 +1,2 @@
+<!-- category: Changed -->
+- **Samples catalog: unified Models demo ([#2239](https://github.com/sceneview/sceneview/issues/2239)).** The retired `multi-model` and `scene-gallery` demos consolidated into the existing flagship `model-viewer` entry with a segmented-button toggle between Single Model (bundled hero viewer with optional Sketchfab "Surprise me" stream), Multi-Model (themed park scene from 4 streamed assets with visibility chips + spin toggle), and Gallery (chip-picked themed Sketchfab models) modes. The `model-viewer` id stays a live registered demo; existing `sceneview://demo/multi-model|scene-gallery` deep links keep working as aliases via `DEMO_ID_ALIASES`.
