--- samples/android-demo/src/main/res/values/strings_demo_scene_gallery.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Strings for the `scene-gallery` demo (registered by
-     `SceneGalleryFragment.kt`). Android's resource merger
-     fans every `res/values/*.xml` file in, so `R.string.demo_scene_gallery_title` etc.
-     remains globally resolvable. Adding a new demo never needs to touch
-     this file or `strings.xml` — see #1870. -->
-<resources>
-    <string name="demo_scene_gallery_title">Scene Gallery</string>
-    <string name="demo_scene_gallery_subtitle">Themed Sketchfab bundles streamed on demand</string>
-    <string name="demo_scene_gallery_credit">by %1$s · CC-BY 4.0</string>
-    <string name="demo_scene_gallery_loading">Streaming model…</string>
-    <string name="demo_scene_gallery_error">Couldn\'t load this model</string>
-    <string name="demo_scene_gallery_retry">Retry</string>
-</resources>
